fix frame allocator

+++ src/util.zig
@@ -1,15 +1,15 @@
 const std = @import("std");
 
 pub const StepIterator = struct {
-    start: u64,
-    end: u64,
-    step_size: u64,
-    next_val: ?u64,
-    pub fn init(start: u64, end: u64, step_size: usize) StepIterator {
+    start: usize,
+    end: usize,
+    step_size: usize,
+    next_val: ?usize,
+    pub fn init(start: usize, end: usize, step_size: usize) StepIterator {
         return StepIterator{ .start = start, .end = end, .step_size = step_size, .next_val = null };
     }
 
-    pub fn next(self: *StepIterator) ?u64 {
+    pub fn next(self: *StepIterator) ?usize {
         if (self.next_val == null) {
             self.next_val = self.start;
             return self.next_val;
@@ -24,6 +24,6 @@ pub const StepIterator = struct {
         return self.next_val;
     }
 };
-pub fn stepBy(start: u64, stop: u64, step: usize) StepIterator {
+pub fn stepBy(start: usize, stop: usize, step: usize) StepIterator {
     return StepIterator.init(start, stop, step);
 }
